Refactoring WindowFunction into coherent structure with AggregateFunction
